feat: firebase storage - auth sync
